refactor: move the shared Modal out of the site-only module
The footer opens one now, and the footer also renders under /admin, where
pulling in (site)'s user context and select is neither wanted nor needed.
