+++ app/(site)/[eventSlug]/comment-likes.tsx
@@ -6,7 +6,7 @@ import { Dialog } from "@headlessui/react";
 
 import { UserContext } from "@/app/(site)/context";
 import { Avatar } from "@/app/(site)/guests/avatar";
-import { Modal } from "@/app/(site)/modals";
+import { Modal } from "@/app/components/modal";
 import type { Comment } from "@/db/repositories/interfaces";
 import { toggleCommentLike } from "./comment-actions";
